Models
results of 3 different models:
Building_1_PV means that just building 1 has a PV system and it is a real case.
Shared_PV means that the real PV system generation of building 1 is shared between all 7 buildings based on their area! It is not real yet!
Designed_PV is not also a real case! Every building has a PV system designed based on the full potential of its area!